packages/bng-monitor/opt/bng-monitor/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets>=12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosqlite==0.20.0
psutil==6.1.0
